+++ client/src/views/ArenaView.tsx
@@ -187,7 +187,7 @@ const ArenaView = () => {
                 value={search}
                 sx={{ mr: 2 }}
               />
-              <Button onClick={searchOpponent} variant="contained">GO !</Button>
+              <Button onClick={searchOpponent} variant="contained">{t('arena.search')}</Button>
             </Box>
           </Grid>
           {!isMd && (